requirements.txt
pyspark==3.5.3
kafka-python==2.0.2
psycopg2-binary==2.9.5
pandas==1.5.3
matplotlib==3.6.3
seaborn==0.12.2
streamlit==1.18.1
python-dotenv==0.21.1
requests==2.28.2